int -> int64
If we define a function that can compare int64 values, then the same function can be used to compare any other integer values (int, int8, int16, int32).

+++ basic/basic.go
@@ -4,7 +4,7 @@ package basic
 // Equaler implements functions for determining if values of basic types are equal.
 type Equaler interface {
 	Bool(bool, bool) bool
-	Int(int, int) bool
+	Int64(int64, int64) bool
 	Float64(float64, float64) bool
 	String(string, string) bool
 	// TODO: add functions for remaining basic types
